Customers controller unit tests and slight adjustment to customers controller

+++ HotelBookingStartupProject/Controllers/CustomersController.cs
@@ -130,7 +130,8 @@ namespace HotelBooking.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult DeleteConfirmed(int id)
         {
-            repository.Remove(id);
+            if (id > 0)
+                repository.Remove(id);
             return RedirectToAction(nameof(Index));
         }
